将数据库操作放到DatabaseAccessObject文件中
